refactor - move API calls and related utils to "api.ts" file
